refactor(test): extract shared soleur-go-runner SDK fixture harness (#3331)
