+++ apps/web-platform/test/soleur-go-runner-chapter-chunked.test.ts
@@ -39,10 +39,7 @@ import {
   afterEach,
 } from "vitest";
 import type {
-  Query,
-  SDKMessage,
   SDKUserMessage,
-  SDKResultMessage,
   SDKAssistantMessage,
 } from "@anthropic-ai/claude-agent-sdk";
 
@@ -88,28 +85,10 @@ import {
   type WorkflowEnd,
 } from "@/server/soleur-go-runner";
 import type { ConversationRouting } from "@/server/conversation-routing";
-
-type Mutable<T> = { -readonly [K in keyof T]: T[K] };
-
-function makeResult(totalCostUsd: number, sessionId = "sess-1"): SDKResultMessage {
-  return {
-    type: "result",
-    subtype: "success",
-    duration_ms: 1,
-    duration_api_ms: 1,
-    is_error: false,
-    num_turns: 1,
-    result: "",
-    stop_reason: "end_turn",
-    total_cost_usd: totalCostUsd,
-    // biome-ignore lint/suspicious/noExplicitAny: minimal SDK fixture
-    usage: { input_tokens: 0, output_tokens: 0 } as any,
-    modelUsage: {},
-    permission_denials: [],
-    uuid: "00000000-0000-0000-0000-0000000000ff" as never,
-    session_id: sessionId,
-  } as SDKResultMessage;
-}
+import {
+  createMockQueryLean as createMockQuery,
+  makeResult,
+} from "./helpers/soleur-go-fixtures";
 
 function makeAssistant(text: string): SDKAssistantMessage {
   return {
@@ -129,75 +108,6 @@ function makeAssistant(text: string): SDKAssistantMessage {
     uuid: "00000000-0000-0000-0000-000000000001" as never,
     session_id: "sess-1",
   } as unknown as SDKAssistantMessage;
-}
-
-function createMockQuery() {
-  let closed = false;
-  const queue: SDKMessage[] = [];
-  let resolveNext: ((r: IteratorResult<SDKMessage>) => void) | null = null;
-  const closeSpy = vi.fn();
-  const iter: AsyncGenerator<SDKMessage, void> = {
-    async next() {
-      if (queue.length > 0) {
-        const v = queue.shift()!;
-        return { value: v, done: false };
-      }
-      if (closed) return { value: undefined, done: true };
-      return new Promise<IteratorResult<SDKMessage>>((r) => {
-        resolveNext = r;
-      });
-    },
-    async return() {
-      closed = true;
-      return { value: undefined, done: true };
-    },
-    async throw(e) {
-      closed = true;
-      throw e;
-    },
-    async [Symbol.asyncDispose]() {
-      closed = true;
-    },
-    [Symbol.asyncIterator]() {
-      return iter;
-    },
-  };
-  function emit(msg: SDKMessage) {
-    if (resolveNext) {
-      const r = resolveNext;
-      resolveNext = null;
-      r({ value: msg, done: false });
-    } else {
-      queue.push(msg);
-    }
-  }
-  function finish() {
-    if (resolveNext) {
-      const r = resolveNext;
-      resolveNext = null;
-      r({ value: undefined, done: true });
-    }
-    closed = true;
-  }
-  const q: Mutable<Partial<Query>> = {
-    ...(iter as unknown as Query),
-    close: () => {
-      closeSpy();
-      finish();
-    },
-    interrupt: vi.fn(async () => {}),
-    setPermissionMode: vi.fn(async () => {}),
-    setModel: vi.fn(async () => {}),
-    setMaxThinkingTokens: vi.fn(async () => {}),
-    applyFlagSettings: vi.fn(async () => {}),
-    // biome-ignore lint/suspicious/noExplicitAny: test stub
-    initializationResult: vi.fn(async () => ({}) as any),
-    supportedCommands: vi.fn(async () => []),
-    supportedModels: vi.fn(async () => []),
-    streamInput: vi.fn(async () => {}),
-    stopTask: vi.fn(async () => {}),
-  };
-  return { query: q as Query, emit, finish, closeSpy };
 }
 
 function makeEvents(): DispatchEvents & {
